+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 - `Visualiser` classmethods to init from RGB/BGR `np.ndarray`
 - Allow `Visualiser.draw_bbox` without a `label`
+- Improve error handling for all methods that accept `xy` or `xyxy` to ensure we have correct `len(xy|xyxy)`
 
 ## 0.4.1 -- 27 Jan 2022
 - Adjust font path, export font with package